+++ lib/pages/home/home_page.dart
@@ -44,7 +44,7 @@ class _HomePageState extends State<HomePage> {
                   icon: const Icon(Icons.add),
                   tooltip: 'Add collection',
                   onPressed: () async {
-                    final collection = await _showEditNameDialog(context);
+                    final collection = await _showEditNameDialog(context, manager);
                     if (collection == null) return;
                     manager.addCollection(collection);
                   },
@@ -281,6 +281,7 @@ class _BodyWidgetState extends State<BodyWidget> {
                   final old = manager.collectionAt(index);
                   final collection = await _showEditNameDialog(
                     context,
+                    manager,
                     oldCollection: old,
                   );
                   if (collection == null) return;
@@ -340,15 +341,21 @@ void _showMessage(BuildContext context, String message) {
 }
 
 Future<Collection?> _showEditNameDialog(
-  BuildContext context, {
+  BuildContext context,
+  HomePageManager manager, {
   Collection? oldCollection,
 }) async {
   final oldName = oldCollection?.name;
-  final versesPerDay = oldCollection?.versesPerDay ?? Collection.defaultVersesPerDay;
   final nameController = TextEditingController(text: oldName);
   StudyStyle studyStyle = oldCollection?.studyStyle ?? StudyStyle.fixedDays;
+
+  // Same number per day
+  final versesPerDay = oldCollection?.versesPerDay ?? Collection.defaultVersesPerDay;
   final versesPerDayController = TextEditingController(text: versesPerDay.toString());
-  final versesPerDayFocusNode = FocusNode();
+
+  // Fixed days
+  final goodDaysController = TextEditingController(text: manager.fixedGoodDays);
+  final easyDaysController = TextEditingController(text: manager.fixedEasyDays);
 
   return showDialog(
     context: context,
@@ -362,7 +369,7 @@ Future<Collection?> _showEditNameDialog(
               children: [
                 TextField(
                   textCapitalization: TextCapitalization.sentences,
-                  autofocus: true,
+                  autofocus: oldName == null,
                   controller: nameController,
                   decoration: const InputDecoration(labelText: 'Name'),
                   onChanged: (value) {
@@ -390,27 +397,35 @@ Future<Collection?> _showEditNameDialog(
                   onChanged: (value) {
                     setState(() {
                       studyStyle = value!;
-                      if (studyStyle == StudyStyle.sameNumberPerDay) {
-                        versesPerDayFocusNode.requestFocus();
-                        versesPerDayController.selection = TextSelection(
-                          baseOffset: 0,
-                          extentOffset: versesPerDayController.text.length,
-                        );
-                      }
                     });
                   },
                   decoration: const InputDecoration(labelText: 'Review style'),
                 ),
-                if (studyStyle == StudyStyle.sameNumberPerDay) const SizedBox(height: 16),
+                const SizedBox(height: 16),
                 if (studyStyle == StudyStyle.sameNumberPerDay)
                   TextField(
                     keyboardType: TextInputType.number,
                     controller: versesPerDayController,
-                    focusNode: versesPerDayFocusNode,
                     decoration: const InputDecoration(
                       labelText: 'Verses per day',
                     ),
                   ),
+                if (studyStyle == StudyStyle.fixedDays) ...[
+                  TextField(
+                    keyboardType: TextInputType.number,
+                    controller: goodDaysController,
+                    decoration: const InputDecoration(
+                      labelText: 'Days for Good',
+                    ),
+                  ),
+                  TextField(
+                    keyboardType: TextInputType.number,
+                    controller: easyDaysController,
+                    decoration: const InputDecoration(
+                      labelText: 'Days for Easy',
+                    ),
+                  ),
+                ]
               ],
             ),
             actions: [
@@ -418,6 +433,8 @@ Future<Collection?> _showEditNameDialog(
                 onPressed: nameController.text.isEmpty
                     ? null
                     : () {
+                        manager.fixedGoodDays = goodDaysController.text;
+                        manager.fixedEasyDays = easyDaysController.text;
                         Navigator.of(context).pop(
                           Collection(
                             id: oldCollection?.id ?? const Uuid().v4(),